refactor(rxCse): make rxCse.cpp a .Call shim like its siblings
seFromSE.cpp is 62 lines and rxToSE.cpp is 30 -- in this cluster the .cpp reads
the argument and hands the result back, and the work lives in a header
(seBatch.h).  rxCse.cpp was 470 lines and the odd one out, so the pipeline
moves to src/rxCseRun.h and the .Call keeps only csFinish, csReadLines and the
entry point.

The phase sequence becomes a table (csPipeline) rather than five nested ifs,
which is what it is: a list of steps, each declining by returning 0.  CS_PHASE
is replaced by csPhaseDone(), so RXCSE_DEBUG=1 still times every step.

No behavior change; the full suite and the byte-exactness fixtures are
unchanged.
